requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
datasets==3.6.0
dill==0.3.8
filelock==3.13.1
fonttools==4.58.1
frozenlist==1.7.0
fsspec==2024.6.1
hf-xet==1.1.5
huggingface-hub==0.33.0
idna==3.10
Jinja2==3.1.4
joblib==1.5.1
kiwisolver==1.4.8
llvmlite==0.44.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.5.0
multiprocess==0.70.16
networkx==3.3
numpy==2.1.2
packaging==25.0
pandas==2.3.0
pillow==11.0.0
propcache==0.3.2
pyarrow==20.0.0
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
rich==14.1.0
rich-argparse==1.7.1
scikit-learn==1.7.1
scipy==1.16.0
six==1.17.0
sympy==1.13.3
threadpoolctl==3.6.0
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.4.0
xxhash==3.5.0
yarl==1.20.1

